+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * [nm] Use API instead of GConf to create new connections
 * [nm] Enable IPv6 on created connections
 * Use GAppInfo to launch applications
+* [bluez] Auto power on adapters; can be disabled in PowerManager settings
 
 ### Bugs fixed