+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### New features
 
-* Support for Python 3.3.
+* Support for Python 3
 * [dhcp] udhcpc (busybox) support
 
 ### Changes